Move PC-SAFT pure to its own header
